requirements.txt
# requirements.txt if you get an error like this: ModuleNotFoundError: No module named 'requests'
# run the command(s): pip install (module name here)
spotipy
python-dotenv
yt-dlp
mutagen
requests
tk
ffmpeg
tkintertable==1.3.3
eyed3
streamlit
spotdl
ipykernel